+++ utils/constant.ts
@@ -1,0 +1,20 @@
+export const themeList = ['dark', 'light', 'radical', 'tokyonight', 'solarized-light', 'ocean-dark']
+export const languages = [
+  {
+    label: '简体中文',
+    value: 'zh-CN',
+  },
+  {
+    label: 'English',
+    value: 'en',
+  },
+]
+export const statisticsList = [
+  'groups',
+  'badges',
+  'games',
+  'screenshots',
+  'guides',
+  'artworks',
+  'reviews',
+]
